+++ plugin-server/src/main/ingestion-queues/anonymous-event-buffer-consumer.ts
@@ -6,7 +6,7 @@ import { KafkaProducerWrapper } from 'utils/db/kafka-producer-wrapper'
 import { KAFKA_BUFFER, KAFKA_EVENTS_DEAD_LETTER_QUEUE } from '../../config/kafka-topics'
 import { JobName } from '../../types'
 import { status } from '../../utils/status'
-import { GraphileWorker } from '../job-queues/concurrent/graphile-worker'
+import { GraphileWorker } from '../jobs/concurrent/graphile-worker'
 import { instrumentEachBatch, setupEventHandlers } from './kafka-queue'
 
 export const startAnonymousEventBufferConsumer = async ({
