+++ plugin-server/src/main/pluginsServer.ts
@@ -24,9 +24,9 @@ import { delay, getPiscinaStats, stalenessCheck } from '../utils/utils'
 import { startAnonymousEventBufferConsumer } from './ingestion-queues/anonymous-event-buffer-consumer'
 import { KafkaQueue } from './ingestion-queues/kafka-queue'
 import { startQueues } from './ingestion-queues/queue'
-import { GraphileWorker } from './job-queues/concurrent/graphile-worker'
-import { startJobQueueConsumer } from './job-queues/job-queue-consumer'
-import { jobQueueMap } from './job-queues/job-queues'
+import { GraphileWorker } from './jobs/concurrent/graphile-worker'
+import { startJobQueueConsumer } from './jobs/job-queue-consumer'
+import { jobQueueMap } from './jobs/job-queues'
 import { createHttpServer } from './services/http-server'
 import { createMmdbServer, performMmdbStalenessCheck, prepareMmdb } from './services/mmdb'
 import { startPluginSchedules } from './services/schedule'
