+++ plugin-server/src/types.ts
@@ -18,7 +18,7 @@ import { Job } from 'node-schedule'
 import { Pool } from 'pg'
 import { VM } from 'vm2'
 
-import { GraphileWorker } from './main/job-queues/concurrent/graphile-worker'
+import { GraphileWorker } from './main/jobs/concurrent/graphile-worker'
 import { ObjectStorage } from './main/services/object_storage'
 import { DB } from './utils/db/db'
 import { KafkaProducerWrapper } from './utils/db/kafka-producer-wrapper'
