+++ plugin-server/src/utils/db/hub.ts
@@ -6,7 +6,7 @@ import { StatsD } from 'hot-shots'
 import Redis from 'ioredis'
 import { Kafka, Partitioners, SASLOptions } from 'kafkajs'
 import { DateTime } from 'luxon'
-import { GraphileWorker } from 'main/job-queues/concurrent/graphile-worker'
+import { GraphileWorker } from 'main/jobs/concurrent/graphile-worker'
 import * as path from 'path'
 import { types as pgTypes } from 'pg'
 import { ConnectionOptions } from 'tls'
@@ -14,7 +14,7 @@ import { ConnectionOptions } from 'tls'
 import { getPluginServerCapabilities } from '../../capabilities'
 import { defaultConfig } from '../../config/config'
 import { KAFKAJS_LOG_LEVEL_MAPPING } from '../../config/constants'
-import { JobQueueManager } from '../../main/job-queues/job-queue-manager'
+import { JobQueueManager } from '../../main/jobs/job-queue-manager'
 import { connectObjectStorage } from '../../main/services/object_storage'
 import { Hub, KafkaSecurityProtocol, PluginServerCapabilities, PluginsServerConfig } from '../../types'
 import { ActionManager } from '../../worker/ingestion/action-manager'
